feat(render): lit skinned material — skinned models shade with the frame's lights
SkinnedMaterial was unlit (texture + fog only), so skinned glTF models
rendered flat next to lit primitives. Extract LitMaterial's light loop,
Blinn-Phong specular, and shadow-map sampling into a shared GLSL snippet
(lighting_glsl / LightingUniforms in light.rs, the FOG_GLSL pattern) and
shade the skinned fragment with it; the skinned vertex shader deforms the
normal by the same joint blend as the position (the
SkinnedNormalDebugMaterial convention). Skinned models now receive shadows
too (they already cast via SkinnedDepthMaterial), and an unlit
Frame.create shades them exactly like LitMaterial (ambient-only when no
lights are supplied).

Regenerated goldens: hello-t2, lighting-t2 (the two scenarios with
skinned models); hello-normals/tangents, primitives, synthwave are
byte-identical.
